services/media/requirements.txt
Flask>=3.0.0,<4.0.0
boto3>=1.28.0
python-dotenv>=1.0.0
PyJWT>=2.8.0
Werkzeug>=3.0.0
pytest>=7.4.0
psycopg2-binary>=2.9.0
